// main/dashboard/src/dom/factory/core/types.ts
export type Child = HTMLElement | SVGElement | string | Instance<HTMLElement>;

export interface Instance<T extends HTMLElement = HTMLElement> {
    el: T;
    mount(parent: Element): Instance<T>;
    addChild(child: Child): Instance<T>;
    addFirst(child: Child): Instance<T>;
    addBefore(child: Child, ref: Node | null): Instance<T>;
    addBatchBefore(children: readonly Child[], ref: Node | null): Instance<T>;
    detach(): Instance<T>;
    destroy(): void;
    setText(text: ReactiveValue<string>): Instance<T>;
    setHTML(html: ReactiveValue<string>): Instance<T>;
    setAttr(name: string, value: ReactiveValue<string | null>): Instance<T>;
    removeAttr(name: string): Instance<T>;
    setChildren(...children: readonly Child[]): Instance<T>;
    clear(): Instance<T>;
    toggleClass(className: string, force?: boolean): Instance<T>;
    addEffect(name: string): Instance<T>;
    removeEffect(name: string): Instance<T>;
    trackDispose(d: Disposable): Instance<T>;
}
// ...
export interface ContextProps {
    context?: string | null;
    meta?: readonly MetaTag[] | null;
}
// ...
export type AttrEntry = readonly [string, string | undefined];